refactor(portal-next): move more filters dialog out of component folder
