+++ gravitee-apim-portal-webui-next/src/app/applications/application/application-tab-logs/more-filters-dialog/more-filters-dialog.component.html
